start new library for VE.Direct protocol
This is the beginning of a library for parsing VE.Direct streaming data
messages. According to the whitepaper for VE.Direct Protocol (v3.29,)
the devices will transmit blocks of data every second. Each block
contains some number of fields delimited by newlines and a final
Checksum field for the block.